requirements.txt
Flask
pymongo
python-dotenv
bcrypt
pandas
scikit-learn
werkzeug
gunicorn
joblib>=1.1.0
numpy>=1.20.0